remove redis and huey from stats.py
a new otree update removes these, so redwood was broken. now it always uses mockredispy to save stats results.
this probably isn't ideal, but if I'm being honest I don't know what stats.py is for. I think it's just something James was using to profile redwood when he originally wrote it, so it's not very relevant anymore.